Make sure that links use the right colour (#1973).
By default, it’s blue while now we use the same colour as any normal
text, which is fine since a link is underlined.

To have blue on a light background is relatively ok, but not on a dark
background. A lighter blue is not great either, hence we use the same
colour of normal text and rely on a link being underlined…