SED-ML L1V4: updated the way we keep track of / retrieve the type of the title of a curve (#2053).
Not sure we want to show a curve's id if it has no name. Indeed, otherwise we would have to do the same for a plot, and there are cases where we might simply not want a "title" for a plot while if a plot was to have no name then we would normally need to show its id, which is not what we want.