GraphPanel widget: improved our tracking/restoring of colours.
Note that it only make sense to retrieve our original colours if we have some graphs and that it's the same number as the number of graph colours that were originally tracked. Indeed, say that that you have a SED-ML file that has a couple of curves including an invalid one (e.g. one of the variables doesn't exist) then it means that only graph colours will be tracked. Then, if we are to correct the SED-ML file outside of OpenCOR. To go back to OpenCOR will trigger us to reload the SED-ML file. If we do so then the number of graphs won't be the same as the number of graph colours, hence the test (not to mention that it prevents a crash because of the out-of-bound index!).