macOS: attempt at handling the "Secure coding is automatically enabled for restorable state!" warning message.
However, this doesn't work. Indeed, Qt overloads the NSApplicationDelegate class (QCocoaApplicationDelegate), so ideally we would overload Qt's interface, but... it's private, so we can't!

To ignore the message in messageHandler() (in src/misc/cliutils.cpp) doesn't work either since it's a pure macOS warning, so not something that Qt can control.

The only option seems to be to upgrade to Qt 6 since it handles that macOS warning, but... we need Qt WebKit and it has yet to be ported to Qt 6. (In fact, I wouldn't expect Qt WebKit to ever be ported to Qt 6. Qt WebKit is already lagging behind big time, so...)